# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amrutam Telemedicine API
  version: 1.0.0
  description: |
    Production-grade REST API for an Ayurveda telemedicine platform.

    ## Authentication
    Protected endpoints require a JWT bearer token:
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.

    ## Idempotency
    `POST /api/v1/appointments` requires an `Idempotency-Key` header (UUID recommended).

    ## Response Envelope
    Success: `{ success, data, meta, requestId, timestamp }`
    Error: `{ success: false, code, message, details, requestId, timestamp }`
  contact:
    name: Amrutam Engineering
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Client-generated UUID for idempotent booking
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, code, message, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    CreateSlotRequest:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    CreateLeaveRequest:
      type: object
      required: [startDate, endDate]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
...
    CompleteConsultationRequest:
      type: object
      properties:
        followUpDate:
          type: string
          format: date
        followUpNotes:
          type: string
    ClinicalNoteRequest:
      type: object
      properties:
        chiefComplaint:
          type: string
        symptoms:
          type: string
        diagnosis:
          type: string
        observations:
          type: string
        doctorNotes:
          type: string
        advice:
          type: string
        followUpInstructions:
          type: string
  responses:
    Unauthorized:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: UNAUTHORIZED
            message: Invalid or missing bearer token
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    Forbidden:
      description: Insufficient role or resource access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Insufficient role or resource access
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: VALIDATION_ERROR
            message: Request validation failed
            details: ['doctorId must be a UUID']
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: NOT_FOUND
            message: Resource not found
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Rate limit exceeded
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: INTERNAL_ERROR
            message: An unexpected error occurred
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
paths:
...
  /doctors:
    get:
      tags: [Doctors]
      summary: Search verified doctors
      description: Public endpoint. Returns VERIFIED doctors supporting online consultations. Results cached 60s via Redis cache-aside.
      security: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
          description: Search name, bio, or specialization
        - name: specialization
          in: query
          schema:
            type: string
          example: ayurveda
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Doctor list
          content:
            application/json:
              example:
                success: true
                data:
                  - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    name: Dr. Ananya Patel
                    yearsOfExperience: 12
                    consultationFee: 500
                    specializations:
                      - name: Ayurveda
                        slug: ayurveda
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /doctors/me/slots:
    post:
      tags: [Doctors]
      summary: Create availability slot
      description: Doctor-only. Creates a new AVAILABLE slot for the authenticated doctor.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSlotRequest'
            example:
              startTime: '2026-07-15T10:00:00.000Z'
              endTime: '2026-07-15T10:30:00.000Z'
      responses:
        '201':
          description: Slot created
          content:
            application/json:
              example:
                success: true
                data:
                  id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  status: AVAILABLE
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /doctors/me/leaves:
    get:
      tags: [Doctors]
      summary: List leave periods
      description: Doctor-only. Returns all leave records for the authenticated doctor.
      responses:
        '200':
          description: Leave list
          content:
            application/json:
              example:
                success: true
                data: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Doctors]
      summary: Create leave period
      description: Doctor-only. Blocks booking during leave dates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeaveRequest'
            example:
              startDate: '2026-08-01'
              endDate: '2026-08-05'
              reason: Annual leave
      responses:
        '201':
          description: Leave created
          content:
            application/json:
              example:
                success: true
                data:
                  id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /doctors/{id}:
    get:
      tags: [Doctors]
      summary: Get doctor profile by ID
      description: Public endpoint. Returns doctor profile with specializations and consultation fee.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Doctor profile
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  name: Dr. Ananya Patel
                  consultationFee: 500
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '404':
          description: Doctor not found
          content:
            application/json:
              example:
                success: false
                code: DOCTOR_NOT_FOUND
                message: Doctor not found
                details: null
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalError'
  /doctors/{id}/slots:
    get:
      tags: [Doctors]
      summary: List available slots for a doctor
      description: Public endpoint. Returns AVAILABLE slots within optional date range. Not cached — must reflect real-time availability.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Available slots
          content:
            application/json:
              example:
                success: true
                data:
                  - id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                    startTime: '2026-07-15T10:00:00.000Z'
                    endTime: '2026-07-15T10:30:00.000Z'
                    status: AVAILABLE
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /consultations/{id}/start:
    post:
      tags: [Consultations]
      summary: Start a consultation
      description: Transitions consultation from SCHEDULED to IN_PROGRESS. Doctor-only.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Consultation started
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: IN_PROGRESS
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid consultation status
          content:
            application/json:
              example:
                success: false
                code: INVALID_CONSULTATION_STATUS
                message: Consultation cannot be started in current status
        '500':
          $ref: '#/components/responses/InternalError'
  /consultations/{id}/complete:
    post:
      tags: [Consultations]
      summary: Complete a consultation
      description: Transitions to COMPLETED, stores follow-up, emits outbox event. Doctor-only.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteConsultationRequest'
            example:
              followUpDate: '2026-08-01'
              followUpNotes: Review in 2 weeks
      responses:
        '200':
          description: Consultation completed
          content:
            application/json:
              example:
                success: true
                data:
                  status: COMPLETED
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid status transition
          content:
            application/json:
              example:
                success: false
                code: INVALID_CONSULTATION_STATUS
                message: Consultation cannot be completed in current status
        '500':
          $ref: '#/components/responses/InternalError'
  /consultations/history:
    get:
      tags: [Consultations]
      summary: Search consultation history
      description: Filterable consultation history scoped to caller role.
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: keyword
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Consultation history
          content:
            application/json:
              example:
                success: true
                data:
                  items: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /consultations/me:
    get:
      tags: [Consultations]
      summary: List my consultations
      description: Returns consultations for authenticated patient or doctor.
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Consultation list
          content:
            application/json:
              example:
                success: true
                data:
                  items: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /consultations/{id}:
    get:
      tags: [Consultations]
      summary: Get consultation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Consultation details
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: SCHEDULED
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /consultations/{id}/notes:
    post:
      tags: [Consultations]
      summary: Create clinical notes
      description: Doctor-only. Creates new append-only clinical note version.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicalNoteRequest'
            example:
              chiefComplaint: Chronic acidity
              diagnosis: Amlapitta
      responses:
        '201':
          description: Notes created
          content:
            application/json:
              example:
                success: true
                data:
                  version: 1
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Consultations]
      summary: Update clinical notes
      description: Doctor-only. Creates new version; previous versions immutable.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicalNoteRequest'
      responses:
        '200':
          description: New note version created
          content:
            application/json:
              example:
                success: true
                data:
                  version: 2
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...